# docs/openapi.json
{
    "components": {"schemas":{"Article":{"properties":{"body":{"example":"Lorem ipsum dolor sit amet, consectetur adipiscing elit.","type":"string"},"created_at":{"example":"2009-11-10T23:00:00Z","format":"RFC3339","type":"string"},"id":{"example":"fe9f6053-8929-4868-be47-f3015c46577b","format":"uuid","type":"string"},"title":{"example":"How to train your AI dragon!","type":"string"},"updated_at":{"example":"2009-11-10T23:30:00Z","format":"RFC3339","type":"string"},"user_id":{"example":"fe9f6053-8929-4868-be47-f3015c46577b","format":"uuid","type":"string"}},"required":["body","created_at","id","title","updated_at","user_id"],"type":"object"},"CreateArticleRequest":{"properties":{"body":{"example":"Lorem ipsum dolor sit amet, consectetur adipiscing elit.","type":"string"},"title":{"example":"How to train your AI dragon!","type":"string"}},"required":["body","title"],"type":"object"},"CreateArticleResponse":{"properties":{"id":{"example":"fe9f6053-8929-4868-be47-f3015c46577b","format":"uuid","type":"string"}},"required":["id"],"type":"object"},"CreateUserRequest":{"properties":{"email":{"example":"example@gmail.com","format":"email","type":"string"},"name":{"example":"username123","type":"string"},"password":{"example":"zaq1@WSXEDC","type":"string"}},"required":["email","name","password"],"type":"object"},"CreateUserResponse":{"properties":{"id":{"example":"fe9f6053-8929-4868-be47-f3015c46577b","type":"string"}},"required":["id"],"type":"object"},"GetAccessTokenRequest":{"properties":{"refresh_token":{"example":"dfr_YWRpQWNrbURTZHZmQ1lhZF9jOWIyZDA2Mg==","type":"string"}},"required":["refresh_token"],"type":"object"},"GetAccessTokenResponse":{"properties":{"access_token":{"example":"dfa_Zk5QTldQc0Vtb2ZOUFduTl9iZjNmMTJlYw==","type":"string"}},"required":["access_token"],"type":"object"},"SignInRequest":{"properties":{"email":{"example":"example@gmail.com","format":"email","type":"string"},"password":{"example":"zaq1@WSXEDC","type":"string"}},"required":["email","password"],"type":"object"},"SignInResponse":{"properties":{"refresh_token":{"example":"dfr_YWRpQWNrbURTZHZmQ1lhZF9jOWIyZDA2Mg==","type":"string"}},"required":["refresh_token"],"type":"object"},"SignOutRequest":{"properties":{"refresh_token":{"example":"dfr_YWRpQWNrbURTZHZmQ1lhZF9jOWIyZDA2Mg==","type":"string"}},"required":["refresh_token"],"type":"object"},"UpdateArticleRequest":{"properties":{"body":{"example":"Lorem ipsum dolor sit amet, consectetur adipiscing elit.","type":"string"},"title":{"example":"How to train your AI dragon!","type":"string"}},"type":"object"},"UpdateUserRequest":{"properties":{"email":{"example":"example@gmail.com","format":"email","type":"string"},"name":{"example":"username123","type":"string"},"password":{"example":"zaq1@WSXEDC","type":"string"}},"type":"object"},"User":{"properties":{"created_at":{"example":"2009-11-10T23:00:00Z","format":"RFC3339","type":"string"},"id":{"example":"fe9f6053-8929-4868-be47-f3015c46577b","format":"uuid","type":"string"},"name":{"example":"John Doe","type":"string"},"updated_at":{"example":"2009-11-10T23:30:00Z","format":"RFC3339","type":"string"}},"required":["created_at","id","name","updated_at"],"type":"object"}},"securitySchemes":{"bearerauth":{"bearerFormat":"JWT","scheme":"bearer","type":"http"}}},
    "info": {"description":"API for interacting with the forum.","title":"Dev-Forum","version":"1.0"},
    "externalDocs": {"description":"","url":""},
    "paths": {"/articles/":{"get":{"operationId":"GetArticles","parameters":[{"description":"items offset","example":60,"in":"query","name":"offset","schema":{"type":"integer"}},{"description":"item limit per page","example":30,"in":"query","name":"limit","schema":{"type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Article"},"type":"array"}}},"description":"OK"},"500":{"description":"An unexpected error occurred."}},"summary":"Get paginated articles.","tags":["articles"]},"post":{"operationId":"CreateArticle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArticleRequest"}}},"description":"Payload","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArticleResponse"}}},"description":"Created"},"400":{"description":"The request body could not be parsed."},"401":{"description":"Authorization token is invalid or missing."},"500":{"description":"An unexpected error occurred."}},"security":[{"bearerauth":[]}],"summary":"Create an article.","tags":["articles"]}},"/articles/{id}":{"delete":{"operationId":"DeleteArticle","parameters":[{"description":"Article ID","example":"fe9f6053-8929-4868-be47-f3015c46577b","in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"204":{"description":"No Content"},"401":{"description":"Authorization token is invalid or missing."},"404":{"description":"Article could not be found."},"500":{"description":"An unexpected error occurred."}},"security":[{"bearerauth":[]}],"summary":"Delete an article.","tags":["articles"]},"get":{"operationId":"GetArticle","parameters":[{"description":"Article ID","example":"fe9f6053-8929-4868-be47-f3015c46577b","in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Article"}}},"description":"OK"},"404":{"description":"Article could not be found."},"500":{"description":"An unexpected error occurred."}},"summary":"Get an article by ID.","tags":["articles"]},"patch":{"description":"Accepts individual fields to update.\nIf no fields are provided then the call is a no-op and status 200 is returned.","operationId":"UpdateArticle","parameters":[{"description":"Article ID","example":"fe9f6053-8929-4868-be47-f3015c46577b","in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateArticleRequest"}}},"description":"Payload","required":true},"responses":{"200":{"description":"OK"},"400":{"description":"The request body could not be parsed or the ID was not given."},"401":{"description":"Authorization token is invalid or missing."},"500":{"description":"An unexpected error occurred."}},"security":[{"bearerauth":[]}],"summary":"Update an article.","tags":["articles"]}},"/auth/get-access-token":{"post":{"operationId":"GetAccessToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAccessTokenRequest"}}},"description":"Payload","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAccessTokenResponse"}}},"description":"OK"},"400":{"description":"Payload could not be parsed or the refresh token is missing/empty."},"500":{"description":"An unexpected error occurred."}},"summary":"\"Get a new access token.\"","tags":["auth"]}},"/auth/sign-in":{"post":{"operationId":"SignIn","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInRequest"}}},"description":"Payload","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponse"}}},"description":"OK"},"400":{"description":"Payload could not be parsed or credentials are missing/empty."},"401":{"description":"Credentials are invalid."},"500":{"description":"An unexpected error occurred."}},"summary":"\"Sign in\"","tags":["auth"]}},"/auth/sign-out":{"post":{"operationId":"SignOut","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignOutRequest"}}},"description":"Payload","required":true},"responses":{"200":{"description":"OK"},"400":{"description":"Payload could not be parsed or the refresh token is missing/empty."},"500":{"description":"An unexpected error occurred."}},"summary":"\"Sign out\"","tags":["auth"]}},"/users/":{"get":{"operationId":"GetUsers","parameters":[{"description":"items offset","example":60,"in":"query","name":"offset","schema":{"type":"integer"}},{"description":"item limit per page","example":30,"in":"query","name":"limit","schema":{"type":"integer"}},{"description":"search filter","example":"name[$eq]=john\u0026email[$eq]=doe@example.com","in":"query","name":"filter","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/User"},"type":"array"}}},"description":"OK"},"500":{"description":"An unexpected error occurred."}},"summary":"Get paginated users.","tags":["users"]},"post":{"operationId":"CreateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"description":"Payload","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserResponse"}}},"description":"Created"},"400":{"description":"The request body could not be parsed."},"500":{"description":"An unexpected error occurred."}},"summary":"Create a user.","tags":["users"]}},"/users/{id}":{"delete":{"operationId":"DeleteUser","parameters":[{"description":"User ID","example":"fe9f6053-8929-4868-be47-f3015c46577b","in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"204":{"description":"No Content"},"401":{"description":"Authorization token is invalid or missing."},"404":{"description":"User could not be found."},"500":{"description":"An unexpected error occurred."}},"security":[{"bearerauth":[]}],"summary":"Delete a user.","tags":["users"]},"get":{"operationId":"GetUser","parameters":[{"description":"User ID","example":"fe9f6053-8929-4868-be47-f3015c46577b","in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"description":"OK"},"404":{"description":"User could not be found."},"500":{"description":"An unexpected error occurred."}},"summary":"Get a user by ID.","tags":["users"]},"patch":{"description":"Accepts individual fields to update.\nIf no fields are provided then the call is a no-op and status 200 is returned.","operationId":"UpdateUser","parameters":[{"description":"User ID","example":"fe9f6053-8929-4868-be47-f3015c46577b","in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"description":"Payload","required":true},"responses":{"200":{"description":"OK"},"400":{"description":"The request body could not be parsed."},"401":{"description":"Authorization token is invalid or missing."},"404":{"description":"User ID is empty or user does not exist."},"500":{"description":"An unexpected error occurred."}},"security":[{"bearerauth":[]}],"summary":"Update a user.","tags":["users"]}}},
    "openapi": "3.1.0"
}